Added support for simple form and changed some views

+++ db/migrate/20220504015938_add_comission_percentage_to_dashboard.rb
@@ -1,5 +1,5 @@
 class AddComissionPercentageToDashboard < ActiveRecord::Migration[6.1]
   def change
-    add_column :dashboards, :comission_percentage, :decimal
+    add_column :dashboards, :comission_percentage, :integer
   end
 end
